openai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "convert-slides"
description = "Tool to translate slides using OpenAI"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "openai==1.26",
    "python-dotenv==1.0.1",
    "beautifulsoup4==4.12.3"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff"
]